Renames
algorithms.py -> abstracts.py
