+++ algorithm_tester/abstracts.py
