+++ package_parsers/__init__.py
@@ -1,6 +1,6 @@
 # from package_algorithms.algorithms_outer import BruteForce_Outer
 # from package_algorithms.algorithms_outer2 import Greedy_Outer
-# from algorithm_tester.algorithms import Algorithm
+# from algorithm_tester.abstracts import Algorithm
 
 # __plugins__ = [BruteForce_Outer, Greedy_Outer]
 # __all__ = [plugin.__name__ for plugin in __plugins__]
